+++ BillSplitter/iOSProjectTemplate/Views & Controllers/BSHeadcountViewController.m
@@ -168,14 +168,14 @@
 
 - (void)stepperValueDidChange:(UIVerticalStepper *)stepper
 {
-	self.textField.text = [NSString stringWithFormat:@"%i", (int)stepper.value];
+	self.textField.text = [NSString stringWithFormat:@"%i", (NSInteger)stepper.value];
 	
 	// Change image based on number
 	// TODO
     
     // Let delegate know
     if (self.delegate && [self.delegate respondsToSelector:@selector(headCountViewController:countChanged:)]) {
-        [self.delegate headCountViewController:self countChanged:(int)stepper.value];
+        [self.delegate headCountViewController:self countChanged:(NSInteger)stepper.value];
     }
      
     // Bounce on change
